Fix crash when for textures many null values are put
I guess it's a corner-case, the schema allows it but in theory the idea
is to put only one null. If someone puts many then it won't crash and
will return valid.

Actually, if the first value is null, then anything after is fine since
it's now ignored.

That was identified with the France Auvergne-Rhône-Alpes-2021 dataset